+++ src/infrastructure/prun-api/prun-api-listener.ts
@@ -1,49 +1,60 @@
-import socketIOMiddleware from './socket-io-middleware';
+import socketIOMiddleware, { Middleware } from './socket-io-middleware';
 import { dispatch } from '@src/infrastructure/prun-api/data/api-messages';
 import { companyContextId } from '@src/infrastructure/prun-api/data/user-data';
 import { startMeasure, stopMeasure } from '@src/utils/performance-measure';
 import { context } from '@src/infrastructure/prun-api/data/screens';
 
-export interface Packet {
+interface Message {
   messageType?: string;
-  payload?: {
-    message: Packet;
-  };
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  payload?: { message: Message } | any;
 }
 
-export function listenPrunApi() {
-  socketIOMiddleware<Packet>(onOpen, payload => {
+const middleware: Middleware<Message> = {
+  onOpen: function () {
+    const storeAction = {
+      type: 'CLIENT_CONNECTION_OPENED',
+      data: undefined,
+    };
+    dispatch(storeAction);
+  },
+  onMessage: message => {
     if (context.value === companyContextId.value || !companyContextId.value || !context.value) {
-      processEvent(payload);
+      processEvent(message);
     }
     return false;
-  });
+  },
+  dispatchClientMessage: undefined,
+};
+
+export function listenPrunApi() {
+  socketIOMiddleware<Message>(middleware);
 }
 
-function onOpen() {
-  const storeAction = {
-    type: 'CLIENT_CONNECTION_OPENED',
-    data: undefined,
-  };
-  dispatch(storeAction);
-}
-
-function processEvent(packet: Packet) {
-  if (!packet || !packet.messageType || !packet.payload) {
+function processEvent(message: Message) {
+  if (!message || !message.messageType || !message.payload) {
     return;
   }
 
-  startMeasure(packet.messageType);
+  startMeasure(message.messageType);
 
-  if (packet.messageType === 'ACTION_COMPLETED') {
-    processEvent(packet.payload.message);
+  if (message.messageType === 'ACTION_COMPLETED') {
+    processEvent(message.payload.message);
   } else {
     const storeAction = {
-      type: packet.messageType,
-      data: packet.payload,
+      type: message.messageType,
+      data: message.payload,
     };
     dispatch(storeAction);
   }
 
   stopMeasure();
+}
+
+export function dispatchClientPrunMessage(message: Message) {
+  if (!middleware.dispatchClientMessage) {
+    return false;
+  }
+  middleware.dispatchClientMessage?.(message);
+  return true;
 }